move `env` & `config` outside server
