+++ package.json
@@ -33,15 +33,15 @@
 	"dependencies": {
 		"@hono/swagger-ui": "~0.4.1",
 		"@hono/zod-openapi": "~0.16.4",
-		"@types/bun": "~1.1.11",
-		"@types/node": "~22.7.7",
+		"@types/bun": "~1.1.12",
+		"@types/node": "~22.8.1",
 		"env-var": "~7.5.0",
-		"hono": "~4.6.5",
+		"hono": "~4.6.7",
 		"typescript": "~5.6.3"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "~1.9.4",
-		"lefthook": "~1.7.22",
+		"lefthook": "~1.8.1",
 		"sort-package-json": "~2.10.1"
 	},
 	"trustedDependencies": [
