+++ src/document/validator.ts
@@ -1,4 +1,4 @@
-import { config } from '../server.ts';
+import { config } from '../config.ts';
 import { errorHandler } from '../server/errorHandler.ts';
 import type { Document } from '../types/Document.ts';
 import { ErrorCode } from '../types/ErrorHandler.ts';
