+++ src/endpoints/v1/access.route.ts
@@ -1,7 +1,7 @@
 import { type OpenAPIHono, createRoute, z } from '@hono/zod-openapi';
+import { config } from '../../config.ts';
 import { compression } from '../../document/compression.ts';
 import { storage } from '../../document/storage.ts';
-import { config } from '../../server.ts';
 import { errorHandler, schema } from '../../server/errorHandler.ts';
 import { ErrorCode } from '../../types/ErrorHandler.ts';
