+++ src/endpoints/v2/access.route.ts
@@ -1,8 +1,8 @@
 import { type OpenAPIHono, createRoute, z } from '@hono/zod-openapi';
 import { storage } from '@x-document/storage.ts';
+import { config } from '../../config.ts';
 import { compression } from '../../document/compression.ts';
 import { validator } from '../../document/validator.ts';
-import { config } from '../../server.ts';
 import { errorHandler, schema } from '../../server/errorHandler.ts';
 import { ErrorCode } from '../../types/ErrorHandler.ts';
