+++ src/endpoints/v2/exists.route.ts
@@ -1,6 +1,6 @@
 import { type OpenAPIHono, createRoute, z } from '@hono/zod-openapi';
+import { config } from '../../config.ts';
 import { validator } from '../../document/validator.ts';
-import { config } from '../../server.ts';
 import { errorHandler, schema } from '../../server/errorHandler.ts';
 import { ErrorCode } from '../../types/ErrorHandler.ts';
