+++ src/endpoints/v2/remove.route.ts
@@ -1,8 +1,8 @@
 import { unlink } from 'node:fs/promises';
 import { type OpenAPIHono, createRoute, z } from '@hono/zod-openapi';
 import { storage } from '@x-document/storage.ts';
+import { config } from '../../config.ts';
 import { validator } from '../../document/validator.ts';
-import { config } from '../../server.ts';
 import { errorHandler, schema } from '../../server/errorHandler.ts';
 import { ErrorCode } from '../../types/ErrorHandler.ts';
