+++ src/server.ts
@@ -1,37 +1,22 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
-import { LogLevels, logger } from '@x-util/logger.ts';
+import { env } from '@x-util/env.ts';
+import { logger } from '@x-util/logger.ts';
 import { serve } from 'bun';
-import { get as envvar } from 'env-var';
 import { cors } from 'hono/cors';
 import { HTTPException } from 'hono/http-exception';
+import { config } from './config.ts';
 import { documentation } from './server/documentation.ts';
 import { endpoints } from './server/endpoints.ts';
 import { errorHandler } from './server/errorHandler.ts';
 import { ErrorCode } from './types/ErrorHandler.ts';
 
-export const env = {
-	port: envvar('PORT').default(4000).asPortNumber(),
-	logLevel: envvar('LOGLEVEL').default(LogLevels.info).asIntPositive(),
-	tls: envvar('TLS').asBoolStrict() ?? true,
-	documentMaxSize: envvar('DOCUMENT_MAXSIZE').default(1024).asIntPositive(),
-	docsEnabled: envvar('DOCS_ENABLED').asBoolStrict() ?? false,
-	docsPath: envvar('DOCS_PATH').default('/docs').asString()
-} as const;
+process.on('SIGTERM', async () => await backend.stop());
 
-export const config = {
-	protocol: env.tls ? 'https://' : 'http://',
-	apiPath: '/api',
-	storagePath: 'storage/',
-	documentNameLengthMin: 2,
-	documentNameLengthMax: 32,
-	documentNameLengthDefault: 8
-} as const;
+logger.set(env.logLevel);
 
 const instance = new OpenAPIHono().basePath(config.apiPath);
 
 export const server = (): typeof instance => {
-	logger.set(env.logLevel);
-
 	instance.use('*', cors());
 
 	instance.onError((err) => {
@@ -57,10 +42,6 @@ export const server = (): typeof instance => {
 };
 
 const backend = serve({
-	port: env.port,
-	fetch: server().fetch
-});
-
-process.on('SIGTERM', async () => {
-	await backend.stop();
+	fetch: server().fetch,
+	port: env.port
 });
