+++ src/server/endpoints.ts
@@ -1,7 +1,7 @@
 import type { OpenAPIHono } from '@hono/zod-openapi';
 import { v1 } from '@x-v1/index.ts';
 import { v2 } from '@x-v2/index.ts';
-import { config } from '../server.ts';
+import { config } from '../config.ts';
 
 export const endpoints = (instance: OpenAPIHono): void => {
 	instance.get('/documents/*', (ctx) => {
